services/python-ai/requirements.txt
pydantic>=2.0,<3
psycopg2-binary>=2.9,<3
python-dotenv>=1.0,<2
httpx>=0.27,<1
pytest>=7.0,<9
